Switch MidiPlayer scheduler
Since we emit the playback state events via an observable now, it is up to the subscribers to choose the correct thread to run their code on, so MidiPlayer is no longer coupled to them in that regard. Now, we can set up the mediaplayer stuff to happen in the background.