Upgrade AS version
This version actually fixes an issue with lower API level devices where there would be a ResourceNotFoundException on start up. I got this before and was able to resolve it by getting rid of some vector images. Now that it is properly fixed, though, I may re-add the vector images back