Catch exception
I think this was happening when I happened to hit the randomize song button as the current song was ending. I think it is a race condition with where I adm doing updates for the new song as well as ending the old song (can probably fix by changing some Rx threading stuff). Anyway, this should at least prevent the app from crashing.